Fix type shortening problem.
Former-commit-id: 80c35e33d47a792113529687d16381d05c107a66

+++ parser/samplerplacement.h
@@ -62,7 +62,7 @@ namespace GMAD
     double aper3;
     double aper4;
     double tilt;
-    double nPoints;
+    int    nPoints;
     
     /// constructor
     SamplerPlacement();
